Rename Command to Scene and reconfigure Component classes
Moving to a component configuration where multiple components
can be added to a Scene. The Scene class replaces the Command
class as the scope of the Command class was too great to be a pure
command. Might reintroduce later to modify scenes.

Note that the siren sound is not working and that tests are
failing (but still building).