Use single interface for components and scene
Drawing and sound components and the scene now use the same
Component interface. This removes a level of obscurity in the
game logic where either a single component or a more complex
scene can be run through the game loop.

Building and not crashing but functionality has currently
regressed for the typewriter spoken scene.