Allow copying of Game object
Copying of the game object is useful for tests that require the
game to be in a particular state that takes some time to reach.
By moving to that state once and then copying, test times can
be reduced.

Copying the game required copying of a number of other objects
that relied on polymorphism with unique_ptr. To avoid the clone
pattern for copying unique_ptr (and multiple methods for different
base classes), a number of unique_ptrs are changed to
polymorphic_value types that support copying out of the box.

The main program is currently working, but the tests have yet to
be updated.